Final version of the Chord DHT
The final version of the Chord DHT. This program gives the user the ability to run more than one queries at the same time (multithreading execution). The user gives an educational institution and an awards threshold for each query. The program executes a lookup, based on the chord protocol. The chord ring contains 512 nodes. A node can leave or join the ring at any time, and the ring's information will not be lost. The program returns (for each query) the scientists who have studied at the given education institution and have more awards than the threshold.